requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --allow-unsafe --strip-extras
#
asgiref==3.8.1
    # via
    #   django
    #   django-cors-headers
django==5.1.4
    # via
    #   django-cors-headers
    #   djangorestframework
    #   djangorestframework-simplejwt
django-cors-headers==4.6.0
    # via -r requirements.in
django-environ==0.11.2
    # via -r requirements.in
djangorestframework==3.15.2
    # via djangorestframework-simplejwt
djangorestframework-simplejwt==5.3.1
    # via -r requirements.in
psycopg2==2.9.10
    # via -r requirements.in
pyjwt==2.10.1
    # via djangorestframework-simplejwt
sqlparse==0.5.3
    # via django
